requirements.txt
numpy  
pandas
scikit-learn
matplotlib
seaborn
pyyaml 
pillow
torch
torchvision
mlflow 
python-dotenv
grad-cam
tqdm 
